fixed bugs updated readme

+++ FaaSr_py/faasr_payload.py
@@ -75,20 +75,17 @@ class FaaSr:
             s3_client = boto3.client(
                 "s3",
                 aws_access_key_id=self.payload_dict["DataStores"][server]["AccessKey"],
-                aws_secret_access_key=self.payload_dict["DataStores"][server][
-                    "SecretKey"
-                ],
+                aws_secret_access_key=self.payload_dict["DataStores"][server]["SecretKey"],
                 region_name=self.payload_dict["DataStores"][server]["Region"],
                 endpoint_url=self.payload_dict["DataStores"][server]["Endpoint"],
             )
             # Use boto3 head bucket to ensure that the bucket exists and that we have acces to it
-            bucket_check = s3_client.head_bucket(
-                Bucket=self.payload_dict["DataStores"][server]["Bucket"]
-            )
-
-            # If bucket check does not return a dict, then the data store is unreachable
-            if not isinstance(bucket_check, dict):
-                error_message = f'{{"s3_check":"S3 server {server} failed with message: Data store server unreachable"}}\n'
+            try:
+                bucket_check = s3_client.head_bucket(
+                    Bucket=self.payload_dict["DataStores"][server]["Bucket"]
+                )
+            except Exception as e:
+                error_message = f'{{"s3_check":"S3 server {server} failed with message: {e}}}\n'
                 print(error_message)
                 sys.exit(1)
 
@@ -170,11 +167,15 @@ class FaaSr:
         # If a predecessor has a rank attribute, then we need to ensure
         # That all concurrent invocations of that function have finished
         for pre_func in pre:
-            if "Rank" in self.payload_dict["FunctionList"][pre_func]:
+            if "Rank" in self.payload_dict["FunctionList"][pre_func] and len(self.payload_dict["FunctionList"][pre_func]["Rank"]) != 0:
                 parts = self.payload_dict["FunctionList"][pre_func]["Rank"].split("/")
-                pre = pre.remove(pre_func)
-            for rank in range(1, parts[1] + 1):
-                pre.append(f"{pre_func}.{rank}")
+                pre.remove(pre_func)
+                if len(parts) != 2:
+                    err_msg = f'{{\"faasr_abort_on_multiple_invocation\": \"Error with rank field in function: {pre_func}\"}}'
+                    print(err_msg)
+                    sys.exit(1)
+                for rank in range(1, int(parts[1]) + 1):
+                    pre.append(f"{pre_func}.{rank}")
 
         # First, we check if all of the other predecessor actions are done
         # To do this, we check a file called func.done in S3, and see if all of the other actions have
@@ -198,7 +199,7 @@ class FaaSr:
             # if it does not exist, then the current function still is waiting for
             # a predecessor and must wait
             if done_file not in s3_object_keys:
-                res_msg = '{"fabort_on_multiple_invocations":"not the last trigger invoked - no flag"}\n'
+                res_msg = '{"faasr_abort_on_multiple_invocations":"not the last trigger invoked - no flag"}\n'
                 print(res_msg)
                 sys.exit(1)
 
@@ -213,37 +214,39 @@ class FaaSr:
         # 4) download the file from S3
 
         # to-do faasr acquire lock
-        self.faasr_acquire()
+        FaaSr_py.faasr_acquire(self)
 
         random_number = random.randint(1, 2**31 - 1)
 
         if not os.path.isdir(id_folder):
-            os.mkdir(id_folder)
+            os.makedirs(id_folder, exist_ok=True)
 
         candidate_path = f"{id_folder}/{self.payload_dict['FunctionInvoke']}.candidate"
 
         # Get all of the objects in S3 with the prefix {id_folder}/{FunctionInvoke}.candidate
         s3_response = s3_client.list_objects_v2(
-            Bucket=s3_log_info["Bucket"], prefix=candidate_path
+            Bucket=s3_log_info["Bucket"], Prefix=candidate_path
         )
-        if len(s3_response["Contents"]) != 0:
+        if 'Contents' in s3_response and len(s3_response["Contents"]) != 0:
+            # Download candidate set
             if os.path.exists(candidate_path):
-                os.remove(candidate_path)
+                os.remove(candidate_path)          
             s3_client.download_file(
                 Bucket=s3_log_info["Bucket"],
                 Key=candidate_path,
                 Filename=candidate_path,
             )
 
-        # Append random number to candidate file
-        with open(candidate_path, "a") as candidate_file:
-            candidate_file.write(random_number + "\n")
+        # Write unique random number to candidate file
+        with open(candidate_path, "a") as cf:
+            cf.write(str(random_number) + "\n")
 
-        # Upload candidate file back to S3
-        s3_client.put_object(
-            Body=candidate_file, Key=candidate_path, Bucket=s3_log_info["Bucket"]
-        )
-
+        with open(candidate_path, "rb") as cf:
+            # Upload candidate file back to S3
+            s3_client.put_object(
+                Body=cf, Key=candidate_path, Bucket=s3_log_info["Bucket"]
+            )
+            
         # Download candidate file to local directory again
         if os.path.exists(candidate_path):
             os.remove(candidate_path)
@@ -252,10 +255,10 @@ class FaaSr:
         )
 
         # Release the lock
-        self.faasr_release()
+        FaaSr_py.faasr_release(self)
 
         # Abort if current function was not the first to write to the candidate set
-        with open(candidate_path, "a") as updated_candidate_file:
+        with open(candidate_path, "r") as updated_candidate_file:
             first_line = updated_candidate_file.readline().strip()
             first_line = int(first_line)
         if random_number != first_line:
@@ -273,7 +276,6 @@ class FaaSr:
             logging_server = self.payload_dict["LoggingDataStore"]
         return logging_server
     
-
     def run_user_function(self, imported_functions):
         """
         This method runs the user's code that was imported
@@ -299,7 +301,7 @@ class FaaSr:
         # Run user function
         try:
             user_function(**user_args)
-        except Exception as e:
+        except Exception:
             nat_err_msg = f'"faasr_run_user_function":Errors in the user function {repr(e)}'
             err_msg = '{"faasr_run_user_function":"Errors in the user function: ' + str(self.payload_dict["FunctionInvoke"]) + ', check the log for the detail "}\n'
             result_2 = FaaSr_py.faasr_log(nat_err_msg)
@@ -395,9 +397,9 @@ class FaaSr:
                 next_server_type = next_compute_server["FaaSType"]
 
                 match(next_server_type):
-                    # to-do: OW and lambda triggers
+                    # to-do: OW and lambda testing
                     case "OpenWhisk":
-                        print("OpenWhisk trigger not implemented")
+                        print("OpenWhisk trigger not implemented (need to test)")
                         # Get ow credentials
                         endpoint = next_compute_server["Endpoint"]
                         api_key = next_compute_server["API.key"]
@@ -407,7 +409,7 @@ class FaaSr:
                         if "SSL" not in next_compute_server or len(next_compute_server["SSL"]) == 0:
                             ssl = True
                         else:
-                            if next_compute_server["SSL"] in ['true', 'True']:
+                            if next_compute_server["SSL"].lower() != 'false': 
                                 ssl = True
                             else:
                                 ssl = False
@@ -429,14 +431,75 @@ class FaaSr:
                             "Content-Type": "application/json"
                         }
 
+                        # Create body for POST
+                        json_payload = json.dumps(self.payload_dict)
+
                         # Issue POST request
-                        response = requests.post(url=url,
-                                                 json=json.dumps(faasr_dict),
-                                                 headers=headers,
-                                                 verify=ssl)
+                        try:
+                            response = requests.post(url=url,
+                                                    auth=(api_key[0], api_key[1]),
+                                                    json=json_payload,
+                                                    headers=headers,
+                                                    verify=ssl)
+                        except Exception as e:
+                            if type(e) == requests.exceptions.ConnectionError:
+                                err_msg = f"{{\"faasr_trigger\": \"OpenWhisk: Error invoking {faasr_dict['FunctionInvoke']} -- connection error\"}}"
+                                print(err_msg)
+                                sys.exit(1)
+                            else:
+                                err_msg = f"{{\"faasr_trigger\": \"OpenWhisk: Error invoking {faasr_dict['FunctionInvoke']} -- see logs\"}}"
+                                nat_err_msg = err_msg = f"{{\"faasr_trigger\": \"OpenWhisk: Error invoking {faasr_dict['FunctionInvoke']} -- error: {e}\"}}"
+                                print(err_msg)
+                                FaaSr_py.faasr_log(nat_err_msg)
+                                sys.exit(1)
+                        
+                        if response.status_code == 200 or response.status_code == 202:
+                            succ_msg = f"{{\"faasr_trigger\":\"OpenWhisk: Succesfully invoked {faasr_dict['FunctionInvoke']}\"}}"
+                            print(succ_msg)
+                            FaaSr_py.faasr_log(succ_msg)
+                        else:
+                            err_msg = f"{{\"faasr_trigger\":\"OpenWhisk: Error invoking {faasr_dict['FunctionInvoke']} -- status code: {response.status_code}\"}}"
+                            print(err_msg)
+                            FaaSr_py.faasr_log(err_msg)
                         break
+
                     case "Lambda":
-                        print("Lamba trigger not implemented")
+                        print("Lamba trigger not implemented (need to test)")
+                        # Create client for invoking lambda function
+                        lambda_client = boto3.client(
+                            "lambda",
+                            aws_access_key_id=next_compute_server["AccessKey"],
+                            aws_secret_access_key=next_compute_server["SecretKey"],
+                            region_name=next_compute_server["Region"],
+                            )
+
+                        # Invoke lambda function
+                        try:
+                            response = lambda_client.invoke(
+                                FunctionName = invoke_next,
+                                InvokeArgs = json.dumps(self.payload_dict),
+                                InvocationType = "Event"
+                                )
+                        except Exception:
+                                err_msg = f"{{\"faasr_trigger\": \"Error invoking function: {self.payload_dict['FunctionInvoke']} -- check API keys\"}}\n"
+                                print(err_msg)
+                                FaaSr_py.faasr_log(err_msg)     
+                                continue                       
+                        
+                        if 'StatusCode' in response and str(response['StatusCode'])[0] == '2':
+                            succ_msg = f"{{\"faasr_trigger\": \"Successfully invoked: {self.payload_dict['FunctionInvoke']}\"}}\n"
+                            print(succ_msg)
+                            FaaSr_py.faasr_log(succ_msg)
+                        else:
+                            try:
+                                err_msg = f"{{\"faasr_trigger\": \"Error invoking function: {self.payload_dict['FunctionInvoke']} -- error: {response['FunctionError']}\"}}\n"
+                                print(err_msg)
+                                FaaSr_py.faasr_log(err_msg)
+                            except Exception:
+                                err_msg = f"{{\"faasr_trigger\": \"Error invoking function: {self.payload_dict['FunctionInvoke']} -- no response from AWS\"}}\n"
+                                print(err_msg)
+                                FaaSr_py.faasr_log(err_msg)
+
                         break
                     case "GitHubActions":
                         # Get env values for GH actions
